Test: Remove OrderTestUtils and inline helpers at call sites
Remove OrderTestUtils test fixture class which had many unused
helper methods. Move the used helpers directly into the test
classes that need them. Duplicate generateSampleOrder in both
WCOrderStoreTest and OrderEntityTest as the only shared helper.
