+++ libs/fluxc-plugin/src/test/java/org/wordpress/android/fluxc/persistence/dao/OrderShipmentProvidersDaoTest.kt
@@ -1,6 +1,8 @@
 package org.wordpress.android.fluxc.persistence.dao
 
 import androidx.test.core.app.ApplicationProvider
+import com.google.gson.JsonElement
+import com.google.gson.JsonParser
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.test.runTest
 import org.assertj.core.api.Assertions.assertThat
@@ -11,9 +13,11 @@ import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
 import org.robolectric.annotation.Config
 import org.wordpress.android.fluxc.UnitTestUtils
+import org.wordpress.android.fluxc.model.LocalOrRemoteId.LocalId
 import org.wordpress.android.fluxc.model.SiteModel
+import org.wordpress.android.fluxc.model.WCOrderShipmentProviderModel
 import org.wordpress.android.fluxc.persistence.DatabaseTestRule
-import org.wordpress.android.fluxc.wc.order.OrderTestUtils
+import kotlin.collections.MutableMap.MutableEntry
 
 @Config(manifest = Config.NONE)
 @RunWith(RobolectricTestRunner::class)
@@ -36,7 +40,7 @@ internal class OrderShipmentProvidersDaoTest {
         val json = UnitTestUtils.getStringFromResourceFile(this.javaClass, "wc/order-shipment-providers.json")
 
         // When
-        val providers = OrderTestUtils.getOrderShipmentProvidersFromJson(json, siteModel.id).apply {
+        val providers = getOrderShipmentProvidersFromJson(json, siteModel.id).apply {
             assertThat(size).isEqualTo(54)
         }
         sut.replaceAll(siteModel.localId(), providers)
@@ -47,12 +51,42 @@ internal class OrderShipmentProvidersDaoTest {
         ).containsExactlyInAnyOrderElementsOf(providers)
 
         // When
-        val updatedProviders = providers.plus(OrderTestUtils.generateOrderShipmentProvider(siteModel.id))
+        val updatedProviders = providers.plus(generateOrderShipmentProvider(siteModel.id))
         sut.replaceAll(siteModel.localId(), updatedProviders)
 
         // Then
         assertThat(
             sut.observeOrderShipmentProviders(siteModel.localId()).first()
         ).containsExactlyInAnyOrderElementsOf(updatedProviders)
+    }
+
+    /* HELPER */
+
+    private fun getOrderShipmentProvidersFromJson(json: String, siteId: Int): List<WCOrderShipmentProviderModel> {
+        val providers = mutableListOf<WCOrderShipmentProviderModel>()
+        val jsonElement = JsonParser().parse(json)
+        jsonElement.asJsonObject.entrySet().forEach { countryEntry: MutableEntry<String, JsonElement> ->
+            countryEntry.value.asJsonObject.entrySet().map { carrierEntry ->
+                carrierEntry?.let { carrier ->
+                    val provider = WCOrderShipmentProviderModel(
+                        localSiteId = LocalId(siteId),
+                        country = countryEntry.key,
+                        carrierName = carrier.key,
+                        carrierLink = carrier.value.asString
+                    )
+                    providers.add(provider)
+                }
+            }
+        }
+        return providers
+    }
+
+    private fun generateOrderShipmentProvider(siteId: Int): WCOrderShipmentProviderModel {
+        return WCOrderShipmentProviderModel(
+            localSiteId = LocalId(siteId),
+            country = "Australia",
+            carrierName = "Amanda Test",
+            carrierLink = "http://google.com"
+        )
     }
 }
